# src/main/java/io/changsoft/fleet_management_mis/rest/VehicleController.java
package io.changsoft.fleet_management_mis.rest;

import io.changsoft.fleet_management_mis.model.VehicleDTO;
import io.changsoft.fleet_management_mis.service.VehicleService;
import java.util.List;
import javax.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;


@RestController
@RequestMapping(value = "/api/vehicles", produces = MediaType.APPLICATION_JSON_VALUE)
public class VehicleController {

    private final VehicleService vehicleService;

    public VehicleController(final VehicleService vehicleService) {
        this.vehicleService = vehicleService;
    }

    @GetMapping
    public ResponseEntity<List<VehicleDTO>> getAllVehicles() {
        return ResponseEntity.ok(vehicleService.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<VehicleDTO> getVehicle(@PathVariable final Long id) {
        return ResponseEntity.ok(vehicleService.get(id));
    }

    @PostMapping
    public ResponseEntity<Long> createVehicle(@RequestBody @Valid final VehicleDTO vehicleDTO) {
        return new ResponseEntity<>(vehicleService.create(vehicleDTO), HttpStatus.CREATED);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Void> updateVehicle(@PathVariable final Long id,
            @RequestBody @Valid final VehicleDTO vehicleDTO) {
        vehicleService.update(id, vehicleDTO);
        return ResponseEntity.ok().build();
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteVehicle(@PathVariable final Long id) {
        vehicleService.delete(id);
        return ResponseEntity.noContent().build();
    }

}